Fix bug in DampedOscillation where it returned infinite values
Parameter Beta was stored in an unsigned.

Turns out (-Beta * t) where Beta is unsigned but t is a double and the
result is a double gives something like (MAX_UINT - a)*t, ie a huge
positive number!

+++ ode_problem.h
@@ -191,7 +191,7 @@ namespace oomph
         return deriv;
       }
 
-      unsigned Beta;
+      double Beta;
       double Omega;
     };
